update1

+++ app/Http/Controllers/Admin/PlaneController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\Plane;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\View;
+use Symfony\Component\HttpFoundation\Response;
 
 class PlaneController extends Controller
 {
@@ -15,8 +17,13 @@ class PlaneController extends Controller
      */
     public function index()
     {
-        $planes=Plane::paginate(20);
-        return view('index.v1.admin.plane.index',compact(['planes']));
+        $planes=Plane::latest('created_at')->paginate(20);
+        if(View::exists('index.v1.admin.plane.index')){
+            return view('index.v1.admin.plane.index',compact(['planes']));
+        }else{
+            abort(Response::HTTP_NOT_FOUND);
+        }
+
     }
 
     /**
@@ -26,7 +33,14 @@ class PlaneController extends Controller
      */
     public function create()
     {
-        return view('index.v1.admin.plane.create');
+
+        if(View::exists('index.v1.admin.plane.create')){
+            return view('index.v1.admin.plane.create');
+
+        }
+        else{
+            abort(Response::HTTP_NOT_FOUND);
+        }
     }
 
     /**
@@ -51,7 +65,7 @@ class PlaneController extends Controller
             $plane->charge=$request->input('charge');
             $plane->save();
             alert()->success('موفقیت آمیز','طرح با موفقیت اضافه شد');
-            return redirect('admin/plane/create');
+            return redirect('admin/plane');
         }
         catch (\Exception $m){
             alert()->warning(' خطا','خطا در ثبت رکورد');
@@ -78,7 +92,19 @@ class PlaneController extends Controller
      */
     public function edit($id)
     {
-        //
+        if(View::exists('index.v1.admin.plane.edit')){
+            $plane=Plane::findorfail($id);
+            if(count(array($plane))>0){
+                return view('index.v1.admin.plane.edit',compact(['plane']));
+            }
+            elseif(count(array($plane))<=0){
+                alert()->error('خطا','کاربری یافت نشد');
+                return redirect('/admin/plane');
+            }
+            elseif(!(View::exists('index.v1.admin.plane.edit'))){
+                abort(Response::HTTP_NOT_FOUND);
+            }
+        }
     }
 
     /**
@@ -90,7 +116,26 @@ class PlaneController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        $this->validate(request(), [
+            'title' => 'required',
+            'description' => 'required',
+            'price' => 'required',
+            'charge' => 'required'
+        ]);
+        try{
+            $plane=Plane::findorfail($id);
+            $plane->title=$request->input('title');
+            $plane->description=$request->input('description');
+            $plane->price=$request->input('price');
+            $plane->charge=$request->input('charge');
+            $plane->save();
+            alert()->success('موفقیت آمیز','طرح با موفقیت ویرایش شد');
+            return redirect('admin/plane');
+        }
+        catch (\Exception $m){
+            alert()->warning(' خطا','خطا در ویرایش رکورد');
+            return redirect('/admin/plane');
+        }
     }
 
     /**
@@ -101,20 +146,20 @@ class PlaneController extends Controller
      */
     public function destroy($id)
     {
-        //
-    }
-    public function delete($id)
-    {
         try{
             $plane=Plane::findorfail($id);
+            if(count(array($plane))>0){
             $plane->delete();
             alert()->success('موفقیت آمیز','طرح با موفقیت حذف شد');
             return redirect('admin/plane');
+            }else{
+                alert()->error('خطا','رکوردی یافت نشد');
+                return redirect('/admin/plane');
+            }
         }
         catch (\Exception $m){
             alert()->erorr(' خطا','خطا در حذف رکورد');
             return redirect('admin/plane');
         }
-
     }
 }
